requirements.txt
aiohttp
aiomysql
loguru
fastapi
pydantic
python-dotenv
uvicorn